backend/requirements.txt
quart>=0.19.0
quart-cors>=0.7.0
raftos>=0.1.1
sqlalchemy>=2.0.0
aiomysql>=0.2.0
pymysql>=1.1.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
